Accept IO objects directly
Accept an IO object for push_file & pull_file.
For push_file, allow the source to be an IO object I.e. check if it responds
to the :read method.
For pull_file, allow the destination to be an IO object I.e. check if it
responds to the :write method.