Push/pull IO objects (#7)
* Accept IO objects directly

Accept an IO object for push_file & pull_file.
For push_file, allow the source to be an IO object I.e. check if it responds
to the :read method.
For pull_file, allow the destination to be an IO object I.e. check if it
responds to the :write method.

* Write the source contents

* Add documentation & tests

Document the changes to pull_file & push_file
Add some tests to cover pulling & pushing a file from an IO object.

* Case is important