add patch profile method
There is to support PATCH lxd api, it should accept a diff hash and call
the lxd#patch. With this method the user of the client does not need to
pass the whole profile. The update_profile method is not very usefull on
its own, because if the user want to update a profile, he needs to GET
it first, parse response, append updates then call update_profile.